Choose highest mag bin in a band as its representative
Band avging attenuated the peak of higher freq bins -- a wrong approach.
How this makes sense? Although, not sure how but here's a rough idea.
Freq comps are orthogonal (e.g. a 200 Hz sinewave does not affect peak
of a 250 Hz sinewave; regardless of phase). Bin with the highest peak
is the peak of a digital signal; so within a band, the bin highest peak
becomes representative of that band.